Extract common quarkus-related device registry classes.
