+++ services/device-registry-base/src/main/java/org/eclipse/hono/deviceregistry/quarkus/KafkaMetricsSupportProducer.java
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.hono.deviceregistry.mongodb.quarkus;
+package org.eclipse.hono.deviceregistry.quarkus;
 
 import java.util.List;
