+++ services/device-registry-mongodb-quarkus/src/main/java/org/eclipse/hono/deviceregistry/mongodb/quarkus/HttpServerFactory.java
@@ -14,7 +14,6 @@
 
 package org.eclipse.hono.deviceregistry.mongodb.quarkus;
 
-import java.util.Objects;
 import java.util.Optional;
 
 import javax.enterprise.context.ApplicationScoped;
@@ -22,20 +21,12 @@ import javax.inject.Inject;
 
 import org.eclipse.hono.deviceregistry.mongodb.config.MongoDbBasedHttpServiceConfigOptions;
 import org.eclipse.hono.deviceregistry.mongodb.config.MongoDbBasedHttpServiceConfigProperties;
+import org.eclipse.hono.deviceregistry.quarkus.AbstractHttpServerFactory;
 import org.eclipse.hono.deviceregistry.server.DeviceRegistryHttpServer;
-import org.eclipse.hono.service.HealthCheckServer;
-import org.eclipse.hono.service.http.HttpEndpoint;
-import org.eclipse.hono.service.management.credentials.CredentialsManagementService;
-import org.eclipse.hono.service.management.credentials.DelegatingCredentialsManagementHttpEndpoint;
-import org.eclipse.hono.service.management.device.DelegatingDeviceManagementHttpEndpoint;
-import org.eclipse.hono.service.management.device.DeviceManagementService;
-import org.eclipse.hono.service.management.tenant.DelegatingTenantManagementHttpEndpoint;
-import org.eclipse.hono.service.management.tenant.TenantManagementService;
+import org.eclipse.hono.service.http.HttpServiceConfigProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.opentracing.Tracer;
-import io.vertx.core.Vertx;
 import io.vertx.ext.auth.mongo.MongoAuthenticationOptions;
 import io.vertx.ext.auth.mongo.impl.DefaultHashStrategy;
 import io.vertx.ext.auth.mongo.impl.MongoAuthenticationImpl;
@@ -47,57 +38,27 @@ import io.vertx.ext.web.handler.BasicAuthHandler;
  *
  */
 @ApplicationScoped
-public class HttpServerFactory {
+public class HttpServerFactory extends AbstractHttpServerFactory {
 
     private static final Logger LOG = LoggerFactory.getLogger(HttpServerFactory.class);
 
     @Inject
-    Vertx vertx;
-
-    @Inject
-    Tracer tracer;
-
-    @Inject
     MongoClient mongoClient;
 
-    @Inject
-    TenantManagementService tenantManagementService;
+    private MongoDbBasedHttpServiceConfigProperties httpServerProperties;
 
     @Inject
-    DeviceManagementService deviceManagementService;
-
-    @Inject
-    CredentialsManagementService credentialsManagementService;
-
-    @Inject
-    HealthCheckServer healthCheckServer;
-
-    private final MongoDbBasedHttpServiceConfigProperties httpServerProperties;
-
-    /**
-     * Creates a new factory.
-     *
-     * @param endpointOptions The HTTP endpoint configuration.
-     */
-    public HttpServerFactory(final MongoDbBasedHttpServiceConfigOptions endpointOptions) {
-        Objects.requireNonNull(endpointOptions);
+    void setHttpServerProperties(final MongoDbBasedHttpServiceConfigOptions endpointOptions) {
         this.httpServerProperties = new MongoDbBasedHttpServiceConfigProperties(endpointOptions);
     }
 
-    /**
-     * Creates a server with an HTTP endpoint exposing Hono's Device Registry Management API.
-     *
-     * @return The server.
-     */
-    public DeviceRegistryHttpServer newServer() {
-        final var server = new DeviceRegistryHttpServer();
-        server.setConfig(httpServerProperties);
-        server.setHealthCheckServer(healthCheckServer);
-        server.setTracer(tracer);
-        server.addEndpoint(tenantHttpEndpoint());
-        server.addEndpoint(deviceHttpEndpoint());
-        server.addEndpoint(credentialsHttpEndpoint());
+    @Override
+    protected final HttpServiceConfigProperties getHttpServerProperties() {
+        return httpServerProperties;
+    }
 
+    @Override
+    protected void customizeServer(final DeviceRegistryHttpServer server) {
         if (httpServerProperties.isAuthenticationRequired()) {
             final var authConfig = httpServerProperties.getAuth();
             LOG.debug("creating AuthenticationHandler guarding access to registry's HTTP endpoint using configuration:{}{}",
@@ -120,52 +81,5 @@ public class HttpServerFactory {
                     mongoAuth,
                     httpServerProperties.getRealm()));
         }
-
-        return server;
-    }
-
-    /**
-     * Creates a new instance of an HTTP protocol handler for the <em>tenants</em> resources
-     * of Hono's Device Registry Management API's.
-     *
-     * @return The handler.
-     */
-    private HttpEndpoint tenantHttpEndpoint() {
-        final var endpoint = new DelegatingTenantManagementHttpEndpoint<TenantManagementService>(
-                vertx,
-                tenantManagementService);
-        endpoint.setConfiguration(httpServerProperties);
-        endpoint.setTracer(tracer);
-        return endpoint;
-    }
-
-    /**
-     * Creates a new instance of an HTTP protocol handler for the <em>devices</em> resources
-     * of Hono's Device Registry Management API's.
-     *
-     * @return The handler.
-     */
-    private HttpEndpoint deviceHttpEndpoint() {
-        final var endpoint = new DelegatingDeviceManagementHttpEndpoint<DeviceManagementService>(
-                vertx,
-                deviceManagementService);
-        endpoint.setConfiguration(httpServerProperties);
-        endpoint.setTracer(tracer);
-        return endpoint;
-    }
-
-    /**
-     * Creates a new instance of an HTTP protocol handler for the <em>credentials</em> resources
-     * of Hono's Device Registry Management API's.
-     *
-     * @return The handler.
-     */
-    private HttpEndpoint credentialsHttpEndpoint() {
-        final var endpoint = new DelegatingCredentialsManagementHttpEndpoint<CredentialsManagementService>(
-                vertx,
-                credentialsManagementService);
-        endpoint.setConfiguration(httpServerProperties);
-        endpoint.setTracer(tracer);
-        return endpoint;
     }
 }
